add current_price to prices controller
